Changes related to: 1.url change

+++ survey_king_web_client/src/layouts/dashboard/index.js
@@ -46,7 +46,7 @@ function Dashboard() {
                 }
             })
             .catch((e) => {
-                if (e.response.status === 403) {
+                if (e.response && e.response.status === 403) {
                     localStorage.clear();
                     navigate('/authentication/sign-in')
                 }
@@ -70,7 +70,7 @@ function Dashboard() {
                 }
             })
             .catch((e) => {
-                if (e.response.status === 403) {
+                if (e.response && e.response.status === 403) {
                     localStorage.clear();
                     navigate('/authentication/sign-in')
                 }
